refactor(Gate): move Extensions.kt to a util package
